Relegate image to "media" folder via github.io
This keeps the blob out of the repository tree, and avoids
the potential risk of github rejecting the image load.

Github has long discouraged the use of serving media directly from the
"raw" or "blob" apis, as it causes a respectable problem for their
services with content retreival.

Using github pages instead fetches from a generated-on-push target,
which is then also replicated via fastly cdn.